added brightness atribute to volume change
added this atribute to handle led switch between listening, or general, and the volume show phase.
Example Situation resolvment: Having the listen LED's sequence set to 0% that I don't have light in my bedroom for sleeping, but when chaning the Volume, no volume level is show. This resolved this issue for me.